Symmetric/Non-symmetric linear logic
I was able to finally see how to define a adjoint logic based on
linear/non-linear logic of Benton that combines commutative linear logic
with non-commutative linear logic.

The answer was in the way that the functor rules for F on the left
and G on the left.  They must preserve the order of the assumptions when
moving from left to right and vice versa.